Netflix-Clone
This Netflix Clone is a web app that replicates Netflix's UI and core features. It allows users to browse and explore movies and TV shows using real-time data fetched from APIs. With a sleek design, it includes trailers, genres, and ratings, providing a dynamic and interactive streaming experience similar to Netflix's original platform.